chore: flatten out structure into a single file
